[nazca] Centralize database
